moved time utility functions to index.util
