+++ tsconfig.json
@@ -19,8 +19,7 @@
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "strict": true,
-    "target": "esnext",
-    "isolatedModules": true
+    "target": "esnext"
   },
   "exclude": [
     "node_modules"
